Do color transform right after de-interleaving a row.
