+++ lib/jpegli/color_transform.h
@@ -6,15 +6,14 @@
 #ifndef LIB_JPEGLI_COLOR_TRANSFORM_H_
 #define LIB_JPEGLI_COLOR_TRANSFORM_H_
 
-#include <stddef.h>
+/* clang-format off */
+#include <stdio.h>
+#include <jpeglib.h>
+/* clang-format on */
 
 #include "lib/jxl/base/compiler_specific.h"
 
 namespace jpegli {
-
-void CMYKToYCCK(float* JXL_RESTRICT row0, float* JXL_RESTRICT row1,
-                float* JXL_RESTRICT row2, float* JXL_RESTRICT row3,
-                size_t xsize);
 
 void YCCKToCMYK(float* JXL_RESTRICT row0, float* JXL_RESTRICT row1,
                 float* JXL_RESTRICT row2, float* JXL_RESTRICT row3,
@@ -23,8 +22,7 @@ void YCCKToCMYK(float* JXL_RESTRICT row0, float* JXL_RESTRICT row1,
 void YCbCrToRGB(float* JXL_RESTRICT row0, float* JXL_RESTRICT row1,
                 float* JXL_RESTRICT row2, size_t xsize);
 
-void RGBToYCbCr(float* JXL_RESTRICT row0, float* JXL_RESTRICT row1,
-                float* JXL_RESTRICT row2, size_t xsize);
+void ChooseColorTransform(j_compress_ptr cinfo);
 
 }  // namespace jpegli
